+++ CHANGES.rst
@@ -52,6 +52,8 @@ Current
       - Enabled for array values.
       - Deprecate argument `sigfigs` [bool] <== `scale` [str]
       - Use `decimals` argument to correctly refer to number of significant figures (when using `sigfigs=True`)
+    - `frexp10()`
+      - Handle the case of `0.0` values.  Set mantissa to 0.0 and exp to np.nan.
     - `spacing()`
       - When creating 'intergers' spacings, allow integers to be subdivided some number of times, specified by an integer value to the `intergers` argument.